feat: add tabs with more (DD-37)
