+++ src/app/features/admin/pages/admin-page/admin-page.component.html
@@ -1,20 +1,1 @@
-<tui-scrollbar class="vertical">
-  <nav tuiTabs vertical="left" class="left" [(activeItemIndex)]="activeItemIndex">
-    <button tuiTab>Stations</button>
-    <button tuiTab>Carriages</button>
-    <button tuiTab>Routes</button>
-  </nav>
-</tui-scrollbar>
-<div class="content">
-  @if (activeItemIndex === 0) {
-    Stations
-  }
-  @if (activeItemIndex === 1) {
-    <dd-carriage-page></dd-carriage-page>
-  }
-  @if (activeItemIndex === 2) {
-    Routes
-  }
-</div>
-
-<!-- <router-outlet /> -->
+<router-outlet />
