+++ src/app/features/admin/pages/admin-page/admin-page.component.ts
@@ -2,12 +2,13 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { TuiTab, TuiTabs } from '@taiga-ui/kit';
 import { TuiScrollbar } from '@taiga-ui/core';
 import { NgClass } from '@angular/common';
+import { RouterLink, RouterOutlet } from '@angular/router';
 import { CarriagePageComponent } from '../carriage-page/carriage-page.component';
 
 @Component({
   selector: 'dd-admin-page',
   standalone: true,
-  imports: [TuiTabs, TuiTab, CarriagePageComponent, TuiScrollbar, NgClass],
+  imports: [TuiTabs, TuiTab, CarriagePageComponent, TuiScrollbar, NgClass, RouterOutlet, RouterLink],
   templateUrl: './admin-page.component.html',
   styleUrl: './admin-page.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
