+++ src/app/features/admin/pages/carriage-page/carriage-page.component.scss
@@ -4,15 +4,15 @@
   display: flex;
   flex-direction: column;
   gap: 2rem;
-  width: 40rem;
+  width: 100%;
 
-  @media #{$tui-mobile} {
-    width: 25rem;
-  }
+  // @media #{$tui-mobile} {
+  //   width: 25rem;
+  // }
 
-  @media #{$tui-mobile-interval} {
-    width: 20rem;
-  }
+  // @media #{$tui-mobile-interval} {
+  //   width: 20rem;
+  // }
 }
 
 .create-btn {
