+++ src/app/features/admin/pages/routes-page/routes-page.component.html
@@ -1,1 +1,17 @@
-<p>routes-page works!</p>
+@let route = currentRoute();
+@if (isEdit()) {
+  @if (route) {
+    <dd-route-form [trainRoute]="route" (save)="onSave()"></dd-route-form>
+  } @else {
+    <dd-route-form (save)="onSave()"></dd-route-form>
+  }
+} @else {
+  <button size="m" tuiButton (click)="onCreate()">Create</button>
+}
+<tui-loader class="loader" [showLoader]="isLoading()">
+  @for (trainRoute of routes(); track trainRoute.id) {
+    <div tuiCardLarge tuiSurface="elevated" class="tui-space_top-4">
+      <dd-route-card [trainRoute]="trainRoute" (edit)="onEdit(trainRoute)" (delete)="onDelete($event)"></dd-route-card>
+    </div>
+  }
+</tui-loader>
