+++ src/app/features/admin/pages/schedule-page/schedule-page.component.scss
@@ -4,6 +4,9 @@
   display: flex;
   flex-direction: column;
   gap: 2rem;
+
+  box-sizing: border-box;
+  width: 100%;
   padding: 2rem;
 
   @media #{$tui-tablet} {
