--- src/app/features/admin/pages/tabs/route-tab/route-tab.component.html
@@ -1,17 +1,0 @@
-@let route = currentRoute();
-@if (isEdit()) {
-  @if (route) {
-    <dd-route-form [trainRoute]="route" (save)="onSave()"></dd-route-form>
-  } @else {
-    <dd-route-form (save)="onSave()"></dd-route-form>
-  }
-} @else {
-  <button size="m" tuiButton (click)="onCreate()">Create</button>
-}
-<tui-loader class="loader" [showLoader]="isLoading()">
-  @for (trainRoute of routes(); track trainRoute.id) {
-    <div tuiCardLarge tuiSurface="elevated" class="tui-space_top-4">
-      <dd-route-card [trainRoute]="trainRoute" (edit)="onEdit(trainRoute)" (delete)="onDelete($event)"></dd-route-card>
-    </div>
-  }
-</tui-loader>
