+++ src/app/features/orders/pages/orders-page/orders-page.component.html
@@ -1,0 +1,1 @@
+<p>orders-page works!</p>
