+++ src/app/features/orders/pages/orders-page/orders-page.component.ts
@@ -1,0 +1,11 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'dd-orders-page',
+  standalone: true,
+  imports: [],
+  templateUrl: './orders-page.component.html',
+  styleUrl: './orders-page.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class OrdersPageComponent {}
